#91: refactored nickname related files for avoiding casing conflics
