+++ src/domain/creator/data/retrieveByNickname.ts
