+++ src/domain/creator/data/retrieveByStartNickname.ts
